# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dehasoft E-Ticaret Çözümü",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Üyelik"
                ],
                "summary": "Giriş Yap",
                "description": "Kullanıcı girişi, eposta ve parola ile",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Kullanıcı bilgilerini girin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mucahit@dehasoft.com.tr"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Başarılı giriş yapıldı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Giriş yapıldı"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Üyelik"
                ],
                "summary": "Kayıt Ol",
                "description": "Kullanıcı kaydı",
                "operationId": "authRegister",
                "requestBody": {
                    "description": "Kullanıcı bilgilerini girin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Mücahit"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Sendinç"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mucahit@dehasoft.com.tr"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kayıt başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Kayıt yapıldı. Eposta adresi doğrulaması gönderildi."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/activate-account": {
            "post": {
                "tags": [
                    "Üyelik"
                ],
                "summary": "Hesap Doğrula",
                "description": "Kullanıcı doğrulama",
                "operationId": "authActiveAccount",
                "requestBody": {
                    "description": "EPosta ile kullanıcıya iletilen tokeni iletin",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "adfaghadfgdagadgdfagadgad"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doğrulama başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Doğrulama yapıldı."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/resend-activation-code": {
            "post": {
                "tags": [
                    "Üyelik"
                ],
                "summary": "Hesap Doğrula Kodu Gönder",
                "description": "Kullanıcı doğrulama kodunu yeniden gönder",
                "operationId": "authResendActivationCode",
                "requestBody": {
                    "description": "EPosta adresini gönderdiğiniz kullanıcıya doğrulama gönderir",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "mucahit@dehasoft.com.tr"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Doğrulama kodu başarı ile gönderildi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Doğrulama kodu gönderildi."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "deha_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "deha_token": {
                "type": "apiKey",
                "description": "Login işleminden dönen tokenType ve tokeni sırasıyla boşluk bırakarak giriniz.",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}